RNG redesign
Exchange the Mersenne Twister PRPG provided by GLib with
xoshiro128**. This allows us to save the state of the PRNG
with the game,thus ensuring the game will always continue
identically when restoring it two times.

As a nice side effect, the new PRNG is significantly faster,
reducing the time required to create the game.

+++ src/monsters.c
@@ -29,6 +29,7 @@
 #include "map.h"
 #include "monsters.h"
 #include "nlarn.h"
+#include "random.h"
 
 /* monster information hiding */
 struct _monster
@@ -1473,7 +1474,7 @@ void monster_player_attack(monster *m, player *p)
         /* make monster size affect weapon damage */
         /* FIXME: handle the vorpal blade */
         dam->amount  = (m->eq_weapon != NULL) ? weapon_damage(m->eq_weapon) : 1
-                        + rand_0n(game_difficulty(nlarn) + 2)
+                        + (int)rand_0n(game_difficulty(nlarn) + 2)
                         + monster_level(m)
                         + 2 * ((monster_size(m) - ESIZE_MEDIUM)) / 25;
     }
